move manifest marshaller into util package
